+++ packages/sdk/src/permissions/permissionEquals.ts
